- started making a navgraph - added a debug mode to show navgraph - a few visual tweaks - altered and increased accuracy of sprite scaling - also fixed minor bug with chainshape in tmsb - cleaned up a few classes - more data validation in level.initialize()

+++ core/src/dev/adamhodgkinson/game/Tile.java
@@ -3,8 +3,8 @@ package dev.adamhodgkinson.game;
 import org.w3c.dom.NamedNodeMap;
 
 public class Tile {
-    private int x;
-    private int y;
+    private short x;
+    private short y;
     String textureName;
     int textureIndex = -1;
 
@@ -16,18 +16,18 @@ public class Tile {
         return textureIndex;
     }
 
-    public int getX() {
+    public short getX() {
         return x;
     }
 
-    public int getY() {
+    public short getY() {
         return y;
     }
 
     public static Tile createTileFromXml(NamedNodeMap attr) {
         Tile t = new Tile();
-        t.x = Integer.parseInt(attr.getNamedItem("x").getNodeValue());
-        t.y = Integer.parseInt(attr.getNamedItem("y").getNodeValue());
+        t.x = Short.parseShort(attr.getNamedItem("x").getNodeValue());
+        t.y = Short.parseShort(attr.getNamedItem("y").getNodeValue());
         t.textureName = attr.getNamedItem("texture").getNodeValue();
         if (attr.getNamedItem("i") != null) {
             t.textureIndex = Integer.parseInt(attr.getNamedItem("i").getNodeValue());
